requirements.txt
locust==2.2.3
faker==4.1.3
PyHamcrest==2.0.2
pytest==6.2.5
